Keep cycle tests out of production module
Move the remaining cycle-detection examples that pushed `cycle.rs` over
the file-size limit into the existing test-only include file after the
rebase onto `origin/main`.

Restore the `rstest` imports and cases needed by the rebased missing-
dependency coverage.
